Refactor header stuff into AboveFold
Fix header colors & paddings
